requirements.txt
setuptools
numpy
scipy
pyopengl
pillow
pyzmq
pyyaml
opencv-python
